examples/pickleball-multi-agent-system/host_agent_adk/pyproject.toml
# This file defines the workspace for the a2a-friend-scheduling project.
# It tells uv that this project is a monorepo containing multiple packages.

[project]
name = "a2a-friend-scheduling"
version = "0.1.0"
description = "A multi-agent system for scheduling pickleball games."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Shared ADK & A2A Dependencies
    "google-adk>=1.2.1",
    "a2a-sdk>=0.2.5",
    "nest-asyncio>=1.6.0",
    "python-dotenv",
    "click",
    "uvicorn",
    "google-generativeai",
    "httpx",
    "streamlit>=1.36",
    "rubix-py>=0.6.0",
    "google-auth-oauthlib>=1.2.3",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.171.0",
    "agent-dna",
]

[tool.uv.sources]
agent-dna = { path = "../../rubix-agent-dna" }
